+++ src/main/java/io/github/smokahs/solpotpie/Sync.java
@@ -23,6 +23,9 @@ public final class Sync {
 	public static final String DIMINISHING_FLOOR_HUNGER = "diminishing_floor_hunger";
 	public static final String DIMINISHING_FLOOR_SATURATION = "diminishing_floor_saturation";
 	public static final String DIMINISHING_RECOVERY_VAL = "diminishing_recovery_val";
+	public static final String FOOD_GROUP_DIVERSITY_THRESHOLD = "food_group_diversity_threshold";
+	public static final String USE_FOOD_GROUPS_AS_WHITELISTS = "use_food_groups_as_whitelists";
+	public static final String NEW_PLAYER_FOODS_EATEN_THRESHOLD = "new_player_foods_eaten_threshold";
 	public static final String SHOULD_FORBIDDEN_COUNT = "should_forbidden_count";
 	public static final String BLACKLIST = "blacklist";
 	public static final String WHITELIST = "whitelist";
@@ -55,6 +58,9 @@ public final class Sync {
 		tag.putInt(DIMINISHING_FLOOR_HUNGER, common.diminishingFloorHunger.get());
 		tag.putDouble(DIMINISHING_FLOOR_SATURATION, common.diminishingFloorSaturation.get());
 		tag.putDouble(DIMINISHING_RECOVERY_VAL, common.diminishingRecoveryVal.get());
+		tag.putInt(FOOD_GROUP_DIVERSITY_THRESHOLD, common.foodGroupDiversityThreshold.get());
+		tag.putBoolean(USE_FOOD_GROUPS_AS_WHITELISTS, common.useFoodGroupsAsWhitelists.get());
+		tag.putInt(NEW_PLAYER_FOODS_EATEN_THRESHOLD, common.newPlayerFoodsEatenThreshold.get());
 
 		tag.putBoolean(SHOULD_FORBIDDEN_COUNT, common.shouldForbiddenCount.get());
 		tag.put(BLACKLIST, strings(common.blacklist.get()));